Add lib folder for utilities functions
