+++ lib/index.ts
@@ -1,0 +1,6 @@
+import {
+    getAreWeHeadlessYetHomePage,
+    getAreWeHeadlessYetTopics,
+} from './fetch';
+
+export { getAreWeHeadlessYetHomePage, getAreWeHeadlessYetTopics };
