+++ pages/index.tsx
@@ -5,6 +5,8 @@ import Head from 'next/head';
 import logo from '../public/images/headless-logo.svg';
 import styles from '../styles/Home.module.scss';
 
+import { getAreWeHeadlessYetHomePage, getAreWeHeadlessYetTopics } from '../lib';
+
 import type { AreWeHeadlessYetHomePage } from '../components/types';
 import type { Topics } from '../components/StreamField/blocks/topics_block';
 
@@ -73,48 +75,9 @@ const Home: NextPage<{ page: AreWeHeadlessYetHomePage; topics: Topics }> = ({
     );
 };
 
-async function fetchHelper(path: string, params: { [key: string]: string }) {
-    let headers = new Headers();
-    if (process.env.INSTANCE === 'staging') {
-        const auth =
-            'Basic ' +
-            Buffer.from(
-                `${process.env.AUTH_USER}:${process.env.AUTH_PASSWORD}`,
-            ).toString('base64');
-        headers.append('Authorization', auth);
-    }
-    return await fetch(
-        `${process.env.BASE_URL}api/v2/pages/${path}?` +
-            new URLSearchParams(params),
-        { headers: headers },
-    );
-}
-
-async function getAreWeHeadlessYetHomePageID() {
-    const response = await fetchHelper('', {
-        type: 'areweheadlessyet.AreWeHeadlessYetHomePage',
-    });
-    const result = await response.json();
-    return result.items[0].id;
-}
-
-async function getAreWeHeadlessYetHomePage() {
-    const HomePageID = await getAreWeHeadlessYetHomePageID();
-    const res = await fetchHelper(HomePageID, {});
-    return await res.json();
-}
-
-async function fetchTopics() {
-    const response = await fetchHelper('', {
-        type: 'areweheadlessyet.AreWeHeadlessYetTopicPage',
-        fields: 'title,status_color,introduction',
-    });
-    return await response.json();
-}
-
 export async function getStaticProps() {
     const page = await getAreWeHeadlessYetHomePage();
-    const topics = await fetchTopics();
+    const topics = await getAreWeHeadlessYetTopics();
 
     return { props: { page: page, topics: topics } };
 }
